+++ AspNetCore.Diagnostics.HealthChecks.sln
@@ -19,8 +19,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{FF4414C2-8
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FunctionalTests", "test\FunctionalTests\FunctionalTests.csproj", "{81A29AE5-5ABC-4D00-AE02-5CFF38E15579}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.IdSvr", "src\HealthChecks.IdSvr\HealthChecks.IdSvr.csproj", "{332F4E33-7D1E-42F6-B509-2950D2B40188}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.Uris", "src\HealthChecks.Uris\HealthChecks.Uris.csproj", "{ED6D5756-ABEB-4EC6-9A09-F1B1EA398F97}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.SqlServer", "src\HealthChecks.SqlServer\HealthChecks.SqlServer.csproj", "{6C9B61AA-BCFC-4230-9407-12B23E522580}"
@@ -118,6 +116,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docker-image", "docker-imag
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.Solr", "src\HealthChecks.Solr\HealthChecks.Solr.csproj", "{6054F41F-6FAA-4E7F-AAE3-5B22228C1468}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthChecks.IdentityServer", "src\HealthChecks.IdentityServer\HealthChecks.IdentityServer.csproj", "{9975B693-EDBA-498E-8798-3B69CE83DCCA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -128,10 +128,6 @@ Global
 		{81A29AE5-5ABC-4D00-AE02-5CFF38E15579}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{81A29AE5-5ABC-4D00-AE02-5CFF38E15579}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{81A29AE5-5ABC-4D00-AE02-5CFF38E15579}.Release|Any CPU.Build.0 = Release|Any CPU
-		{332F4E33-7D1E-42F6-B509-2950D2B40188}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{332F4E33-7D1E-42F6-B509-2950D2B40188}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{332F4E33-7D1E-42F6-B509-2950D2B40188}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{332F4E33-7D1E-42F6-B509-2950D2B40188}.Release|Any CPU.Build.0 = Release|Any CPU
 		{ED6D5756-ABEB-4EC6-9A09-F1B1EA398F97}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{ED6D5756-ABEB-4EC6-9A09-F1B1EA398F97}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ED6D5756-ABEB-4EC6-9A09-F1B1EA398F97}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -292,6 +288,10 @@ Global
 		{252BB504-B7CB-4581-8CD8-D7398CAA16F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{252BB504-B7CB-4581-8CD8-D7398CAA16F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{252BB504-B7CB-4581-8CD8-D7398CAA16F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Release|Any CPU.Build.0 = Release|Any CPU
 		{692313D3-E947-494A-83B7-754E2FFAF348}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{692313D3-E947-494A-83B7-754E2FFAF348}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{692313D3-E947-494A-83B7-754E2FFAF348}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -304,17 +304,16 @@ Global
 		{6054F41F-6FAA-4E7F-AAE3-5B22228C1468}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6054F41F-6FAA-4E7F-AAE3-5B22228C1468}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6054F41F-6FAA-4E7F-AAE3-5B22228C1468}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9975B693-EDBA-498E-8798-3B69CE83DCCA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9975B693-EDBA-498E-8798-3B69CE83DCCA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9975B693-EDBA-498E-8798-3B69CE83DCCA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9975B693-EDBA-498E-8798-3B69CE83DCCA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{81A29AE5-5ABC-4D00-AE02-5CFF38E15579} = {FF4414C2-8863-4ADA-8A1D-4B9F25C361FE}
-		{332F4E33-7D1E-42F6-B509-2950D2B40188} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{ED6D5756-ABEB-4EC6-9A09-F1B1EA398F97} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{6C9B61AA-BCFC-4230-9407-12B23E522580} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{65E0BF0F-B3B6-4DFD-A3E3-F91FE02DA7D3} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
@@ -355,10 +354,11 @@ Global
 		{18F9E412-646D-4751-9751-30AA7A0233DF} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{B526834E-9392-4749-BAB2-7DF579F8F418} = {092533AB-7505-4EDC-8932-D40BF575D0D2}
 		{252BB504-B7CB-4581-8CD8-D7398CAA16F5} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
+		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{692313D3-E947-494A-83B7-754E2FFAF348} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 		{737E4FD6-EA77-4608-A20F-767557FE3190} = {95119F6F-87C8-45B8-8D95-61736FBEBEDE}
 		{6054F41F-6FAA-4E7F-AAE3-5B22228C1468} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
-		{AC711D43-DD3D-4C7C-BF31-1EAADEF50E10} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
+		{9975B693-EDBA-498E-8798-3B69CE83DCCA} = {2A3FD988-2BB8-43CF-B3A2-B70E648259D4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2B8C62A1-11B6-469F-874C-A02443256568}
