--- test/FunctionalTests/HealthChecks.IdSvr/IdSvrHealthCheckTests.cs
@@ -1,79 +1,0 @@
-﻿using FluentAssertions;
-using FunctionalTests.Base;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Diagnostics.HealthChecks;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.TestHost;
-using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Net;
-using System.Threading.Tasks;
-using Xunit;
-
-namespace FunctionalTests.HealthChecks.IdSvr
-{
-    [Collection("execution")]
-    public class idsvr_healthcheck_should
-    {
-        private readonly ExecutionFixture _fixture;
-
-        public idsvr_healthcheck_should(ExecutionFixture fixture)
-        {
-            _fixture = fixture ?? throw new ArgumentNullException(nameof(fixture));
-        }
-
-        [SkipOnAppVeyor]
-        public async Task be_unhealthy_if_idsvr_is_unavailable()
-        {
-            var webHostBuilder = new WebHostBuilder()
-                .UseStartup<DefaultStartup>()
-                .ConfigureServices(services =>
-                {
-                    services.AddHealthChecks()
-                    .AddIdentityServer(new Uri("http://localhost:7777"), tags: new string[] { "idsvr" });
-                })
-                .Configure(app =>
-                {
-                    app.UseHealthChecks("/health", new HealthCheckOptions()
-                    {
-                        Predicate = r => r.Tags.Contains("idsvr")
-                    });
-                });
-
-            var server = new TestServer(webHostBuilder);
-
-            var response = await server.CreateRequest($"/health")
-                .GetAsync();
-
-            response.StatusCode
-                .Should().Be(HttpStatusCode.ServiceUnavailable);
-        }
-
-        [SkipOnAppVeyor]
-        public async Task be_healthy_if_idsvr_is_available()
-        {
-            var webHostBuilder = new WebHostBuilder()
-                .UseStartup<DefaultStartup>()
-                .ConfigureServices(services =>
-                {
-                    services.AddHealthChecks()
-                    .AddIdentityServer(new Uri("http://localhost:8888"), tags: new string[] { "idsvr" });
-                })
-                .Configure(app=>
-                {
-                    app.UseHealthChecks("/health", new HealthCheckOptions()
-                    {
-                        Predicate = r => r.Tags.Contains("idsvr")
-                    });
-                });
-
-            var server = new TestServer(webHostBuilder);
-
-            var response = await server.CreateRequest($"/health")
-                .GetAsync();
-
-            response.StatusCode
-                .Should().Be(HttpStatusCode.OK);
-        }
-    }
-}
