+++ test/UnitTests/DependencyInjection/IdentityServer/IdentityServerUnitTests.cs
@@ -1,0 +1,47 @@
+﻿using FluentAssertions;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Microsoft.Extensions.Options;
+using System;
+using System.Linq;
+using HealthChecks.IdentityServer;
+using Xunit;
+
+namespace UnitTests.HealthChecks.DependencyInjection.IdentityServer
+{
+    public class identityserver_registration_should
+    {
+        [Fact]
+        public void add_health_check_when_properly_configured()
+        {
+            var services = new ServiceCollection();
+            services.AddHealthChecks()
+                .AddIdentityServer(new Uri("http://myidsvr"));
+
+            var serviceProvider = services.BuildServiceProvider();
+            var options = serviceProvider.GetService<IOptions<HealthCheckServiceOptions>>();
+
+            var registration = options.Value.Registrations.First();
+            var check = registration.Factory(serviceProvider);
+
+            registration.Name.Should().Be("identityserver");
+            check.GetType().Should().Be(typeof(IdentityServerHealthCheck));
+        }
+        [Fact]
+        public void add_named_health_check_when_properly_configured()
+        {
+            var services = new ServiceCollection();
+            services.AddHealthChecks()
+                .AddIdentityServer(new Uri("http://myidsvr"), name: "my-idsvr-group");
+
+            var serviceProvider = services.BuildServiceProvider();
+            var options = serviceProvider.GetService<IOptions<HealthCheckServiceOptions>>();
+
+            var registration = options.Value.Registrations.First();
+            var check = registration.Factory(serviceProvider);
+
+            registration.Name.Should().Be("my-idsvr-group");
+            check.GetType().Should().Be(typeof(IdentityServerHealthCheck));
+        }
+    }
+}
